moved patient intake and example runner
