+++ src/examples/foundational/06-listen-and-respond.py
@@ -6,7 +6,7 @@ from dailyai.services.daily_transport_service import DailyTransportService
 from dailyai.services.azure_ai_services import AzureLLMService, AzureTTSService
 from dailyai.services.ai_services import FrameLogger
 from dailyai.pipeline.aggregators import LLMAssistantContextAggregator, LLMUserContextAggregator
-from examples.foundational.support.runner import configure
+from examples.support.runner import configure
 
 
 async def main(room_url: str, token):
