+++ src/examples/foundational/06a-image-sync.py
@@ -16,7 +16,7 @@ from dailyai.services.ai_services import AIService
 from dailyai.pipeline.aggregators import LLMAssistantContextAggregator, LLMUserContextAggregator
 from dailyai.services.fal_ai_services import FalImageGenService
 
-from examples.foundational.support.runner import configure
+from examples.support.runner import configure
 
 
 class ImageSyncAggregator(AIService):
